- removed warnings on msvc compiler - on msvc: loader is now built with orxonox, not as a separate library

+++ code/branches/FICN/src/orxonox/core/Language.cc
@@ -244,7 +244,7 @@ namespace orxonox
             // Check if the line is empty
             if (lineString.compare("") != 0)
             {
-                unsigned int pos = (unsigned int)lineString.find('=');
+                unsigned int pos = lineString.find('=');
 
                 // Check if the length is at least 3 and if there's an entry before and behind the =
                 if (pos > 0 && pos < (lineString.size() - 1) && lineString.size() >= 3)
@@ -287,7 +287,7 @@ namespace orxonox
             // Check if the line is empty
             if (lineString.compare("") != 0)
             {
-                unsigned int pos = (unsigned int)lineString.find('=');
+                unsigned int pos = lineString.find('=');
 
                 // Check if the length is at least 3 and if there's an entry before and behind the =
                 if (pos > 0 && pos < (lineString.size() - 1) && lineString.size() >= 3)
