+++ code/branches/FICN/src/orxonox/objects/Model.cc
@@ -68,6 +68,6 @@ namespace orxonox
     }
 
     void Model::registerAllVariables(){
-      registerVar(&meshSrc_, (int)meshSrc_.length() + 1, network::STRING);
+      registerVar(&meshSrc_, meshSrc_.length() + 1, network::STRING);
     }
 }
